mausjaeger/requirements.txt
RPi.GPIO
picamera